refactor(services): split encoders into focused modules
Break the 880-line encoders.ts monolith into a directory of focused
modules (base64/url/jwt/hash + index). Public surface preserved via
index.ts re-exports so callers continue to import from
'@/lib/services/encoders' unchanged.

Each codec now lives in its own file (~100-360 LOC) with a single
responsibility. CryptoJS import is now isolated to hash.ts.
